+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add overdue search v2 feature flag
 - Implement query to fetch village and patient names in the facility
 - Add custom text field component with chip input
+- Display search suggestions after loading in overdue search screen
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342